LayoutTests/fast/webgpu/nocrash/fuzz-298315-expected.txt

CONSOLE MESSAGE: 0,0
CONSOLE MESSAGE: 414141,0,0,0,66000000,0,0,0
CONSOLE MESSAGE: no validation error
